requirements.txt
Flask
flask-cors
Gunicorn
pyplate-hte